LayoutTests/imported/w3c/web-platform-tests/dom/events/scrolling/wheel-event-transactions-target-resize_include=passive-false-expected.txt


FAIL Resize the initial target and use a passive:false listener assert_equals: Wheel event at `1` does not match the expected target expected Element node <div id="initialTarget" class="spacer" style="background:... but got Element node <div id="firstRootSpacer" class="spacer" style="backgroun...
